refactor(Calendar): convert to hooks #2
