--- src/providers/Auth/index.js
@@ -1,1 +1,0 @@
-export { AuthcConsumer, AuthProvider } from "providers/Auth/Auth";
